+++ packages/saml/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/saml
 
+## 0.1.47
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.17.0
+
 ## 0.1.46
 
 ### Patch Changes